Correctly use of `std::numeric_limits<T>::epsilon`
Previously, `T` == `double`, which did not match expected behavior from
the original implementation, where `T` == `float`.

+++ trac_ik_lib/src/nlopt_ik.cpp
@@ -66,7 +66,7 @@ double minfuncDQ(const std::vector<double>& x, std::vector<double>& grad, void* 
 
   std::vector<double> vals(x);
 
-  double jump = std::numeric_limits<double>::epsilon();
+  double jump = std::numeric_limits<float>::epsilon();
   double result[1];
   c->cartDQError(vals, result);
 
@@ -100,7 +100,7 @@ double minfuncSumSquared(const std::vector<double>& x, std::vector<double>& grad
 
   std::vector<double> vals(x);
 
-  double jump = std::numeric_limits<double>::epsilon();
+  double jump = std::numeric_limits<float>::epsilon();
   double result[1];
   c->cartSumSquaredError(vals, result);
 
@@ -134,7 +134,7 @@ double minfuncL2(const std::vector<double>& x, std::vector<double>& grad, void* 
 
   std::vector<double> vals(x);
 
-  double jump = std::numeric_limits<double>::epsilon();
+  double jump = std::numeric_limits<float>::epsilon();
   double result[1];
   c->cartL2NormError(vals, result);
 
@@ -173,7 +173,7 @@ void constrainfuncm(uint m, double* result, uint n, const double* x, double* gra
     vals[i] = x[i];
   }
 
-  double jump = std::numeric_limits<double>::epsilon();
+  double jump = std::numeric_limits<float>::epsilon();
 
   c->cartSumSquaredError(vals, result);
 
@@ -238,7 +238,7 @@ NLOPT_IK::NLOPT_IK(const KDL::Chain& _chain, const KDL::JntArray& _q_min, const 
 
   assert(types.size() == lb.size());
 
-  std::vector<double> tolerance(1, std::numeric_limits<double>::epsilon());
+  std::vector<double> tolerance(1, std::numeric_limits<float>::epsilon());
   opt.set_xtol_abs(tolerance[0]);
